Undo modifs to data.py

+++ tomo_challenge/data.py
@@ -75,7 +75,7 @@ def download_data():
             filename = f'{f}.hdf5'
             progress = MyProgressBar()
             urlretrieve(f'{url_root_buzzard}/{filename}', f'data_buzzard/{filename}', reporthook=progress)
-            
+
 
 def load_mags(filename, bands, errors=False):
 
@@ -94,10 +94,10 @@ def load_mags(filename, bands, errors=False):
     with h5py.File(filename, 'r') as f:
         # load all bands
         for b in bands:
-            data[b] = f[f'{b}_mag'][:].astype('float32')
+            data[b] = f[f'{b}_mag'][:]
 
             if errors:
-                data[f'{b}_err'] = f[f'{b}_mag_err'][:].astype('float32')
+                data[f'{b}_err'] = f[f'{b}_mag_err'][:]
 
 
     # Set undetected objects to mag 30 +/- 30
@@ -134,7 +134,7 @@ def dict_to_array(data, bands, errors=False, colors=False):
     if errors:
         ncol *= 2
 
-    arr = np.empty((ncol, nobj), dtype='float32')
+    arr = np.empty((ncol, nobj))
     i = 0
     for b in bands:
         arr[i] = data[b]
@@ -181,7 +181,7 @@ def load_data(filename, bands, colors=False, errors=False, array=False):
 def load_redshift(filename):
     """Load a redshift column from a training or validation file"""
     f = h5py.File(filename, 'r')
-    z = f['redshift_true'][:].astype('float32')
+    z = f['redshift_true'][:]
     f.close()
     return z
